Reorganized classes according to the last pull request's comments.
